+++ wanikani/templates/wanikani/base.html
@@ -18,7 +18,7 @@
     <div class="wrapper">
       <div class="navigation">
           <a href="{% url 'index' %}">Home</a>
-          <a href="{% url 'test' %}">Review</a>
+          <a href="{% url 'session' %}">Review</a>
           <a href="{% url 'characters' %}">Characters</a>
           <a href="{% url 'logout' %}">Logout</a>
       </div>
