+++ wanikani/urls.py
@@ -1,21 +1,4 @@
-"""wanikani URL Configuration
-
-The `urlpatterns` list routes URLs to views. For more information please see:
-    https://docs.djangoproject.com/en/2.0/topics/http/urls/
-Examples:
-Function views
-    1. Add an import:  from my_app import views
-    2. Add a URL to urlpatterns:  path('', views.home, name='home')
-Class-based views
-    1. Add an import:  from other_app.views import Home
-    2. Add a URL to urlpatterns:  path('', Home.as_view(), name='home')
-Including another URLconf
-    1. Import the include() function: from django.urls import include, path
-    2. Add a URL to urlpatterns:  path('blog/', include('blog.urls'))
-"""
-
 from django.contrib import admin
-from django.contrib.auth import views as auth_views
 from django.urls import include, path
 
 from . import views
@@ -23,9 +6,10 @@ from . import views
 urlpatterns = [
     path('admin/', admin.site.urls),
     path('', views.index, name='index'),
-    path('test/', views.test, name='test'),
     path('characters/', views.characters, name='characters'),
     path('character/<str:character>/', views.character, name='character'),
-    path('characters_list/', views.test_characters, name='test_characters'),
+    path('session/', views.session, name='session'),
+
     path('registration/', include('wanikani.registration.urls')),
+    path('session/', include('wanikani.session.urls')),
 ]
